docker_shared_folder/UNLPImage/requirements.txt
PySimpleGUI==4.60.5
Pillow==9.5.0